+++ frontend/src/pages/Login.jsx
@@ -1,40 +1,15 @@
 import viteLogo from "/vite.svg";
-import { useState } from "react";
-import { loginUser } from "../api/authService";
-import Cookies from "js-cookie";
+import useLogin from "../hooks/useLogin";
 
 function Login() {
-  const [username, setUsername] = useState("");
-  const [password, setPassword] = useState("");
-  const [alert, setAlert] = useState("");
-  const [token, setToken] = useState(Cookies.get("authToken"));
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-    const submittedUser = {
-      username,
-      password,
-    };
-    const response = await loginUser(submittedUser);
-
-    if (response.success !== false) {
-      setAlert("");
-      Cookies.set("authToken", response.data.token, {
-        expires: 7,
-        secure: true,
-        sameSite: "strict",
-      });
-      setToken(response.data.token);
-    } else {
-      console.log(response);
-      setAlert("Login failed! " + response.error);
-    }
-  };
-
-  const handleLogout = () => {
-    Cookies.remove("authToken");
-    setToken(null);
-  };
+  const {
+    handleSubmit,
+    handleLogout,
+    setUsername,
+    setPassword,
+    loggedIn,
+    alertMessage,
+  } = useLogin();
 
   return (
     <>
@@ -44,7 +19,7 @@ function Login() {
         </a>
       </div>
       <h1>REAL TALK</h1>
-      {token ? (
+      {loggedIn ? (
         <form onSubmit={handleLogout}>
           <p>You are logged in</p>
           <button style={{ width: "96px" }}>Logout</button>
@@ -95,10 +70,10 @@ function Login() {
               margin: "1em",
               minHeight: "2em",
               borderRadius: "5px",
-              visibility: alert ? "visible" : "hidden",
+              visibility: alertMessage ? "visible" : "hidden",
             }}
           >
-            {alert}
+            {alertMessage}
           </div>
 
           <button style={{ width: "96px", marginTop: "1em" }}>Login</button>
